Atualização

+++ app/Http/Controllers/Admin/PlanController.php
@@ -52,10 +52,20 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-      $plan = $this->repository->where('url', $url)->first();
+      $plan = $this->repository
+                    ->with('details')
+                    ->where('url', $url)
+                    ->first();
       
       if (!$plan) 
        return redirect()->back();
+
+      if ($plan->details->count() > 0){
+         return redirect()
+                ->back()
+                ->with('error', 'Existe detalhes vinculados a esse plano, portanto não pode deletar');
+      }
+
 
       $plan->delete();
